+++ test/integration/connectors/test_milvus.py
@@ -173,12 +173,10 @@ def test_precheck_fails_on_nonexistent_collection(collection: str):
 def test_precheck_fails_on_nonexisting_db(collection: str):
     uploader = MilvusUploader(
         connection_config=MilvusConnectionConfig(uri=DB_URI),
-        upload_config=MilvusUploaderConfig(
-            db_name="nonexisting_db", collection_name=collection
-        ),
+        upload_config=MilvusUploaderConfig(db_name="nonexisting_db", collection_name=collection),
     )
     with pytest.raises(
-            DestinationConnectionError,
-            match=f"database not found",
+        DestinationConnectionError,
+        match="database not found",
     ):
         uploader.precheck()